Moved public methods to top
Moved public methods to the top of each file to more easily see the
public interfaces of objects. MapLoader was also extracted from Map.
