+++ routeFinder/model/.gitignore
@@ -5,3 +5,4 @@
 /Street.class
 /AStarSearch.class
 /SearchPoint.class
+/MapLoader.class
